refactor: Move public components to src/lib/seam/components
