+++ src/lib/index.ts
@@ -1,2 +1,3 @@
+export * from './seam/components/index.js'
 export * from './seam/index.js'
-export * from './ui/index.js'
+export * from './seam/SeamProvider.js'
