+++ src/lib/seam/components/AccessCodeDetails/AccessCodeDetails.stories.tsx
@@ -1,7 +1,7 @@
 import { Button, Dialog } from '@mui/material'
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { AccessCodeDetails } from 'lib/ui/AccessCodeDetails/AccessCodeDetails.js'
+import { AccessCodeDetails } from 'lib/seam/components/AccessCodeDetails/AccessCodeDetails.js'
 import { useToggle } from 'lib/ui/use-toggle.js'
 
 /**
