+++ src/lib/seam/components/AccessCodeDetails/AccessCodeDetails.tsx
@@ -4,8 +4,8 @@ import type { AccessCode } from 'seamapi'
 
 import { useAccessCode } from 'lib/index.js'
 
-import { AccessCodeDevice } from 'lib/ui/AccessCodeDetails/AccessCodeDevice.js'
-import { DeviceDetails } from 'lib/ui/DeviceDetails/DeviceDetails.js'
+import { AccessCodeDevice } from 'lib/seam/components/AccessCodeDetails/AccessCodeDevice.js'
+import { DeviceDetails } from 'lib/seam/components/DeviceDetails/DeviceDetails.js'
 import { ContentHeader } from 'lib/ui/layout/ContentHeader.js'
 import { useIsDateInPast } from 'lib/ui/use-is-date-in-past.js'
