+++ src/lib/seam/components/AccessCodeTable/AccessCodeTable.stories.tsx
@@ -1,26 +1,25 @@
-import { Close as CloseIcon } from '@mui/icons-material'
-import { Button, Dialog, DialogActions, IconButton } from '@mui/material'
+import { Button, Dialog } from '@mui/material'
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { DeviceDetails } from 'lib/ui/DeviceDetails/DeviceDetails.js'
+import { AccessCodeTable } from 'lib/seam/components/AccessCodeTable/AccessCodeTable.js'
 import { useToggle } from 'lib/ui/use-toggle.js'
 
 /**
- * These stories showcase the device details.
+ * These stories showcase the access code table.
  */
-const meta: Meta<typeof DeviceDetails> = {
-  title: 'Example/DeviceDetails',
-  component: DeviceDetails,
+const meta: Meta<typeof AccessCodeTable> = {
+  title: 'Example/AccessCodeTable',
+  component: AccessCodeTable,
   tags: ['autodocs'],
 }
 
 export default meta
 
-type Story = StoryObj<typeof DeviceDetails>
+type Story = StoryObj<typeof AccessCodeTable>
 
 export const Content: Story = {
   render: (props, { globals }) => (
-    <DeviceDetails
+    <AccessCodeTable
       {...props}
       deviceId={props.deviceId ?? globals['deviceId']}
     />
@@ -35,26 +34,12 @@ export const InsideModal: Story = {
       <>
         <Button onClick={toggleOpen}>Open Modal</Button>
         <Dialog open={open} fullWidth maxWidth='sm' onClose={toggleOpen}>
-          <IconButton
-            sx={{
-              position: 'absolute',
-              top: '4px',
-              right: '8px',
-            }}
-          >
-            <CloseIcon />
-          </IconButton>
           <div className='seam-components'>
-            <DeviceDetails
+            <AccessCodeTable
               {...props}
               deviceId={props.deviceId ?? globals['deviceId']}
             />
           </div>
-          <DialogActions
-            sx={{ justifyContent: 'center', marginBottom: '16px' }}
-          >
-            <Button variant='outlined'>Done</Button>
-          </DialogActions>
         </Dialog>
       </>
     )
