+++ src/lib/seam/components/DeviceDetails/DeviceModel.tsx
