+++ src/lib/seam/components/DeviceTable/DeviceHealthBar.tsx
