+++ src/lib/seam/components/DeviceTable/DeviceRow.tsx
@@ -7,7 +7,7 @@ import { BatteryStatus } from 'lib/ui/device/BatteryStatus.js'
 import { DeviceImage } from 'lib/ui/device/DeviceImage.js'
 import { LockStatus } from 'lib/ui/device/LockStatus.js'
 import { OnlineStatus } from 'lib/ui/device/OnlineStatus.js'
-import { getDeviceModel } from 'lib/ui/DeviceDetails/DeviceModel.js'
+import { getDeviceModel } from 'lib/seam/components/DeviceDetails/DeviceModel.js'
 import { TableCell } from 'lib/ui/Table/TableCell.js'
 import { TableRow } from 'lib/ui/Table/TableRow.js'
 import { Title } from 'lib/ui/typography/Title.js'
