+++ src/lib/seam/components/SupportedDeviceTable/SupportedDeviceContent.tsx
@@ -2,8 +2,8 @@ import { useQuery } from '@tanstack/react-query'
 import { useState } from 'react'
 
 import { Button } from 'lib/ui/Button.js'
-import { SupportedDeviceFilterArea } from 'lib/ui/SupportedDeviceTable/SupportedDeviceFilterArea.js'
-import type { DeviceModel, Filters } from 'lib/ui/SupportedDeviceTable/types.js'
+import { SupportedDeviceFilterArea } from 'lib/seam/components/SupportedDeviceTable/SupportedDeviceFilterArea.js'
+import type { DeviceModel, Filters } from 'lib/seam/components/SupportedDeviceTable/types.js'
 
 import { SupportedDeviceHeader } from './SupportedDeviceHeader.js'
 import { SupportedDeviceRow } from './SupportedDeviceRow.js'
