+++ src/lib/seam/components/SupportedDeviceTable/SupportedDeviceHeader.tsx
