+++ src/lib/seam/components/SupportedDeviceTable/types.ts
